+++ src/hrtf/kdtree.h
@@ -50,18 +50,18 @@ void kd_clear(struct kdtree *tree);
 void kd_data_destructor(struct kdtree *tree, void (*destr)(void*));
 
 /* insert a node, specifying its position, and optional data */
-int kd_insert(struct kdtree *tree, const double *pos, void *data);
+int kd_insert(struct kdtree *tree, const float *pos, void *data);
 int kd_insertf(struct kdtree *tree, const float *pos, void *data);
-int kd_insert3(struct kdtree *tree, double x, double y, double z, void *data);
+int kd_insert3(struct kdtree *tree, float x, float y, float z, void *data);
 int kd_insert3f(struct kdtree *tree, float x, float y, float z, void *data);
 
 /* Find the nearest node from a given point.
  *
  * This function returns a pointer to a result set with at most one element.
  */
-struct kdres *kd_nearest(struct kdtree *tree, const double *pos);
+struct kdres *kd_nearest(struct kdtree *tree, const float *pos);
 struct kdres *kd_nearestf(struct kdtree *tree, const float *pos);
-struct kdres *kd_nearest3(struct kdtree *tree, double x, double y, double z);
+struct kdres *kd_nearest3(struct kdtree *tree, float x, float y, float z);
 struct kdres *kd_nearest3f(struct kdtree *tree, float x, float y, float z);
 
 /* Find the N nearest nodes from a given point.
@@ -73,9 +73,9 @@ struct kdres *kd_nearest3f(struct kdtree *tree, float x, float y, float z);
  * The result set must be deallocated with kd_res_free after use.
  */
 /*
- struct kdres *kd_nearest_n(struct kdtree *tree, const double *pos, int num);
+ struct kdres *kd_nearest_n(struct kdtree *tree, const float *pos, int num);
  struct kdres *kd_nearest_nf(struct kdtree *tree, const float *pos, int num);
- struct kdres *kd_nearest_n3(struct kdtree *tree, double x, double y, double z);
+ struct kdres *kd_nearest_n3(struct kdtree *tree, float x, float y, float z);
  struct kdres *kd_nearest_n3f(struct kdtree *tree, float x, float y, float z);
  */
 
@@ -87,12 +87,12 @@ struct kdres *kd_nearest3f(struct kdtree *tree, float x, float y, float z);
  * a valid result set is always returned which may contain 0 or more elements.
  * The result set must be deallocated with kd_res_free after use.
  */
-struct kdres *kd_nearest_range(struct kdtree *tree, const double *pos,
-		double range);
+struct kdres *kd_nearest_range(struct kdtree *tree, const float *pos,
+		float range);
 struct kdres *kd_nearest_rangef(struct kdtree *tree, const float *pos,
 		float range);
-struct kdres *kd_nearest_range3(struct kdtree *tree, double x, double y,
-		double z, double range);
+struct kdres *kd_nearest_range3(struct kdtree *tree, float x, float y,
+		float z, float range);
 struct kdres *kd_nearest_range3f(struct kdtree *tree, float x, float y, float z,
 		float range);
 
@@ -116,9 +116,9 @@ int kd_res_next(struct kdres *set);
 /* returns the data pointer (can be null) of the current result set item
  * and optionally sets its position to the pointers(s) if not null.
  */
-void *kd_res_item(struct kdres *set, double *pos);
+void *kd_res_item(struct kdres *set, float *pos);
 void *kd_res_itemf(struct kdres *set, float *pos);
-void *kd_res_item3(struct kdres *set, double *x, double *y, double *z);
+void *kd_res_item3(struct kdres *set, float *x, float *y, float *z);
 void *kd_res_item3f(struct kdres *set, float *x, float *y, float *z);
 
 /* equivalent to kd_res_item(set, 0) */
