+++ src/hrtf/loudness.c
@@ -15,9 +15,9 @@
 #include "mysofa.h"
 #include "tools.h"
 
-double mysofa_loudness(struct MYSOFA_HRTF *hrtf) {
-	double c[3], factor;
-	double min = DBL_MAX;
+float mysofa_loudness(struct MYSOFA_HRTF *hrtf) {
+	float c[3], factor;
+	float min = FLT_MAX;
 	int radius = 0;
 	int i, index;
 	int cartesian = verifyAttribute(hrtf->SourcePosition.attributes, "Type",
