+++ src/hrtf/tools.h
@@ -15,20 +15,20 @@ int changeAttribute(struct MYSOFA_ATTRIBUTE *attr, char *name, char *value,
 		char *newvalue);
 int verifyAttribute(struct MYSOFA_ATTRIBUTE *attr, char *name, char *value);
 
-void convertCartesianToSpherical(double *values, int elements);
-void convertSphericalToCartesian(double *values, int elements);
+void convertCartesianToSpherical(float *values, int elements);
+void convertSphericalToCartesian(float *values, int elements);
 int fequalsf(float a, float b);
-int fequals(double a, double b);
-double radius(double *cartesian);
-double distance(double *cartesian1, double *cartesian2);
+int fequals(float a, float b);
+float radius(float *cartesian);
+float distance(float *cartesian1, float *cartesian2);
 
-void copyToFloat(float *out, double *in, int size);
-void copyFromFloat(double *out, float *in, int size);
+void copyToFloat(float *out, float *in, int size);
+void copyFromFloat(float *out, float *in, int size);
 
-void copyArrayWeighted(double *dst, double *src, int size, double w);
-void addArrayWeighted(double *dst, double *src, int size, double w);
-void scaleArray(double *dst, int size, double w);
-double loudness(double *in, int size);
+void copyArrayWeighted(float *dst, float *src, int size, float w);
+void addArrayWeighted(float *dst, float *src, int size, float w);
+void scaleArray(float *dst, int size, float w);
+float loudness(float *in, int size);
 
 void nsearch(const void *key, const void *base, size_t num, size_t size,
 		int (*cmp)(const void *key, const void *elt), int *lower, int *higher);
